Revert "fix(IVYPORTAL-20673): Case Details - Custom Widgets - Unable to show processes that are not on the start list"
This reverts commit 9d69c507a476130aea87ab4a24fb3db21fa0dfc5.
