+++ AxonIvyPortal/portal/src/ch/ivy/addon/portalkit/bean/AbstractConfigurableContentBean.java
@@ -19,8 +19,7 @@ import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.Strings;
 
-import com.axonivy.portal.components.service.impl.ProcessService;
-import com.axonivy.portal.components.util.ProcessStartUtils;
+import com.axonivy.portal.components.publicapi.ProcessStartAPI;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.JsonMappingException;
@@ -158,8 +157,7 @@ public abstract class AbstractConfigurableContentBean<T extends AbstractConfigur
       if (widget instanceof CustomWidget) {
         CustomWidget customWidget = (CustomWidget) widget;
         if (StringUtils.isNotBlank(customWidget.getData().getProcessPath())) {
-          String url = ProcessStartUtils.findStartableLinkByUserFriendlyRequestPath(customWidget.getData().getProcessPath(),
-              ProcessService.getInstance().findAllProcesses());
+          String url = ProcessStartAPI.findStartableLinkByUserFriendlyRequestPath(customWidget.getData().getProcessPath());
           customWidget.getData().setUrl(url);
         }
       }
